requirements.txt
-i https://pypi.org/simple
alembic==1.7.6
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.3; python_version >= '3'
importlib-metadata==4.11.1; python_version < '3.9'
importlib-resources==5.4.0; python_version < '3.9'
mako==1.1.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.1.0; python_version >= '3.7'
oauthlib==3.2.0; python_version >= '3.6'
psycopg2==2.9.3
python-twitter==3.5
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy==1.4.31
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zipp==3.7.0; python_version < '3.10'
